Abstracted parsing function into its own class
